Refactor module layout of command classes
