+++ autogpt/app.py
@@ -4,7 +4,8 @@ from typing import Dict, List, Union
 
 from autogpt.agent.agent import Agent
 from autogpt.agent.agent_manager import AgentManager
-from autogpt.commands.command import CommandRegistry, command
+from autogpt.command_decorator import command
+from autogpt.models.command_registry import CommandRegistry
 from autogpt.commands.web_requests import scrape_links, scrape_text
 from autogpt.processing.text import summarize_text
 from autogpt.speech import say_text
