+++ autogpt/commands/analyze_code.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 
 from autogpt.agent.agent import Agent
-from autogpt.commands.command import command
+from autogpt.command_decorator import command
 from autogpt.llm.utils import call_ai_function
