+++ autogpt/commands/audio_text.py
@@ -4,7 +4,7 @@ import json
 import requests
 
 from autogpt.agent.agent import Agent
-from autogpt.commands.command import command
+from autogpt.command_decorator import command
 
 
 @command(
