+++ autogpt/commands/git_operations.py
@@ -3,7 +3,7 @@
 from git.repo import Repo
 
 from autogpt.agent.agent import Agent
-from autogpt.commands.command import command
+from autogpt.command_decorator import command
 from autogpt.url_utils.validators import validate_url
