+++ autogpt/commands/task_statuses.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from typing import NoReturn
 
 from autogpt.agent.agent import Agent
-from autogpt.commands.command import command
+from autogpt.command_decorator import command
 from autogpt.logs import logger
